requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
blis==1.3.3
cachetools==7.1.1
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.24.0
comm==0.2.3
confection==1.3.3
cymem==2.0.13
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Events==0.5
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.29.0
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.5.0
grpcio==1.80.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.15
installer==0.7.0
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
jedi==0.20.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.3.0
langchain-classic==1.0.7
langchain-community==0.4.1
langchain-core==1.4.0
langchain-experimental==0.4.1
langchain-milvus==0.3.3
langchain-ollama==1.1.0
langchain-openai==1.2.1
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langdetect==1.0.9
langgraph==1.2.0
langgraph-checkpoint==4.1.0
langgraph-prebuilt==1.1.0
langgraph-sdk==0.3.14
langsmith==0.8.4
llvmlite==0.47.0
lxml==6.1.0
markdown-it-py==4.2.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.2
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numba==0.65.1
numpy==2.4.4
olefile==0.47
ollama==0.6.2
openai==2.36.0
opensearch-protobufs==1.2.0
opensearch-py==3.2.0
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pandas==3.0.3
parso==0.8.7
pexpect==4.9.0
platformdirs==4.9.6
preshed==3.0.13
prompt_toolkit==3.0.52
propcache==0.5.2
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
Pygments==2.20.0
pymilvus==2.6.13
pypdf==6.11.0
pypdfium2==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.4.20
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.5
regex==2026.5.9
requests==2.34.1
requests-toolbelt==1.0.0
rich==15.0.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.1
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
stack-data==0.6.3
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.15.0
typer==0.25.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.22.28
unstructured-client==0.44.0
urllib3==2.7.0
uuid_utils==0.15.0
wasabi==1.1.3
wcwidth==0.7.0
weasel==1.0.0
webencodings==0.5.1
wrapt==2.1.2
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
